+++ dmr/openapi/objects/parameter.py
@@ -15,7 +15,7 @@ class ParameterMetadata:
     """Describes metadata for a single operation parameter."""
 
     description: str | None = None
-    deprecated: bool = False
+    deprecated: bool | None = None
     allow_empty_value: bool | None = None
     style: str | None = None
     explode: bool | None = None
@@ -32,4 +32,4 @@ class Parameter(ParameterMetadata):
     param_in: Annotated[str, Field(alias='in')]
     schema: 'Reference | Schema | None' = None
     content: dict[str, 'MediaType'] | None = None
-    required: bool = False
+    required: bool | None = None