+++ dmr/validation/endpoint_metadata.py
@@ -5,7 +5,6 @@ import warnings
 from collections.abc import (
     Callable,
     ItemsView,
-    KeysView,
     Sequence,
     Set,
 )
@@ -16,6 +15,8 @@ from typing import (
     Any,
     ClassVar,
     Final,
+    Literal,
+    ParamSpec,
     TypeVar,
     assert_never,
 )
@@ -94,6 +95,169 @@ _HTTP_METHODS_WITHOUT_BODY: Final = frozenset((
 
 _PluggableT = TypeVar('_PluggableT', bound=Parser | Renderer)
 _ItemT = TypeVar('_ItemT')
+_HttpSpecCallback = ParamSpec('_HttpSpecCallback')
+
+
+@dataclasses.dataclass(slots=True, frozen=True, kw_only=True)
+class _HttpSpecValidator:  # noqa: WPS214
+    """Collects all http spec validation callbacks."""
+
+    #: 1xx responses, 204, 205, and 304 must not have a body. RFC 9110.
+    _no_response_body_statuses: ClassVar[frozenset[HTTPStatus]] = frozenset((
+        HTTPStatus.NO_CONTENT,
+        HTTPStatus.RESET_CONTENT,
+        HTTPStatus.NOT_MODIFIED,
+    ))
+
+    metadata: EndpointMetadata
+
+    def validate(
+        self,
+        responses: list[ResponseSpec],
+    ) -> None:
+        self._check_http_spec_rule(
+            rule=HttpSpec.header_name_syntax,
+            callback=self._check_http_syntax,
+            responses=responses,
+            field_type='cookie',
+        )
+
+        self._check_http_spec_rule(
+            rule=HttpSpec.header_name_syntax,
+            callback=self._check_http_syntax,
+            responses=responses,
+            field_type='header',
+        )
+
+        self._check_http_spec_rule(
+            rule=HttpSpec.header_name_server_managed,
+            callback=self._check_header_name_server_managed,
+            responses=responses,
+        )
+
+        self._check_http_spec_rule(
+            rule=HttpSpec.empty_response_body,
+            callback=self._check_empty_response_body,
+            responses=responses,
+        )
+
+    def _check_http_spec_rule(
+        self,
+        rule: HttpSpec,
+        callback: Callable[_HttpSpecCallback, None],
+        *args: _HttpSpecCallback.args,
+        **kwargs: _HttpSpecCallback.kwargs,
+    ) -> None:
+        if rule not in self.metadata.no_validate_http_spec:
+            callback(*args, **kwargs)
+
+    def _check_empty_response_body(
+        self,
+        responses: list[ResponseSpec],
+    ) -> None:
+        endpoint_name = self.metadata.endpoint_name
+        # For several http status codes and successful HEAD responses,
+        # no response body is allowed.
+        # If you specify a return annotation other than None,
+        # an EndpointMetadataError will be raised.
+        for response in responses:
+            if not is_safe_subclass(response.return_type, NoneType) and (
+                response.status_code < HTTPStatus.OK
+                or response.status_code in self._no_response_body_statuses
+                or (
+                    stringify(self.metadata.method).upper() == HTTPMethod.HEAD
+                    and response.status_code < HTTPStatus.BAD_REQUEST
+                )
+            ):
+                raise EndpointMetadataError(
+                    f'Can only return `None` not {response.return_type} '
+                    f'from an endpoint {endpoint_name!r} '
+                    f'with status code {response.status_code}',
+                )
+
+    def _check_header_name_server_managed(
+        self,
+        responses: list[ResponseSpec],
+    ) -> None:
+        endpoint_name = self.metadata.endpoint_name
+        for response in responses:
+            if not response.headers:
+                continue
+
+            forbidden_header = self._get_forbidden_header(
+                response.headers.items(),
+            )
+
+            if forbidden_header:
+                raise EndpointMetadataError(
+                    f'Header {forbidden_header!r} is not allowed in responses '
+                    f'from endpoint {endpoint_name!r}.',
+                )
+
+    def _check_http_syntax(
+        self,
+        responses: list[ResponseSpec],
+        field_type: Literal['cookie', 'header'],
+    ) -> None:
+        names = []
+
+        modification = self.metadata.modification
+
+        if modification:
+            names = self._get_http_field_names(
+                modification,
+                field_type,
+            )
+
+        for response in responses:
+            response_names = self._get_http_field_names(
+                response,
+                field_type,
+            )
+            names.extend(response_names)
+
+        invalid_name = self._check_invalid_tokens(names)
+
+        if invalid_name:
+            raise EndpointMetadataError(
+                f'{field_type.capitalize()} name {invalid_name!r} '
+                f'is not following http spec.',
+            )
+
+    def _get_http_field_names(
+        self,
+        resource: ResponseSpec | ResponseModification,
+        field_type: Literal['cookie', 'header'],
+    ) -> list[str]:
+        attribute = getattr(resource, f'{field_type}s')
+
+        if not attribute:
+            return []
+
+        return list(attribute.keys())
+
+    def _check_invalid_tokens(
+        self,
+        names: list[str],
+    ) -> str | None:
+        for name in names:
+            if not _ALLOWED_TOKENS_PATTERN.match(name):
+                return name
+
+        return None
+
+    def _get_forbidden_header(
+        self,
+        response_headers: ItemsView[str, HeaderSpec],
+    ) -> str | None:
+
+        for header_name, header in response_headers:
+            if (
+                header_name.lower() in _FORBIDDEN_RESPONSE_HEADERS
+                and not header.skip_validation
+            ):
+                return header_name
+        return None
 
 
 @dataclasses.dataclass(slots=True, frozen=True, kw_only=True)
@@ -101,12 +265,7 @@ class _ResponseListValidator:  # noqa: WPS214
     """Validates responses metadata."""
 
     metadata: EndpointMetadata
-    #: 1xx responses, 204, 205, and 304 must not have a body. RFC 9110.
-    _no_response_body_statuses: ClassVar[frozenset[HTTPStatus]] = frozenset((
-        HTTPStatus.NO_CONTENT,
-        HTTPStatus.RESET_CONTENT,
-        HTTPStatus.NOT_MODIFIED,
-    ))
+    http_spec_validator: ClassVar[type[_HttpSpecValidator]] = _HttpSpecValidator
 
     def __call__(
         self,
@@ -178,187 +337,17 @@ class _ResponseListValidator:  # noqa: WPS214
         responses: list[ResponseSpec],
     ) -> None:
         """Validate that we don't violate HTTP spec."""
-        if (
-            HttpSpec.empty_response_body
-            not in self.metadata.no_validate_http_spec
-        ):
-            self._check_empty_response_body(responses)
-        if (
-            HttpSpec.header_name_server_managed
-            not in self.metadata.no_validate_http_spec
-        ):
-            self._check_header_name_server_managed(responses)
-
-        if (
-            HttpSpec.header_name_syntax
-            not in self.metadata.no_validate_http_spec
-        ):
-            self._check_modification_header_syntax()
-            self._check_responses_header_syntax(responses)
-
-        if (
-            HttpSpec.cookie_name_syntax
-            not in self.metadata.no_validate_http_spec
-        ):
-            self._check_modification_cookie_syntax()
-            self._check_response_cookie_syntax(responses)
-
-        # TODO: add more checks
-
-    def _check_empty_response_body(
-        self,
-        responses: list[ResponseSpec],
-    ) -> None:
-        endpoint_name = self.metadata.endpoint_name
-        # For several http status codes and successful HEAD responses,
-        # no response body is allowed.
-        # If you specify a return annotation other than None,
-        # an EndpointMetadataError will be raised.
-        for response in responses:
-            if not is_safe_subclass(response.return_type, NoneType) and (
-                response.status_code < HTTPStatus.OK
-                or response.status_code in self._no_response_body_statuses
-                or (
-                    stringify(self.metadata.method).upper() == HTTPMethod.HEAD
-                    and response.status_code < HTTPStatus.BAD_REQUEST
-                )
-            ):
-                raise EndpointMetadataError(
-                    f'Can only return `None` not {response.return_type} '
-                    f'from an endpoint {endpoint_name!r} '
-                    f'with status code {response.status_code}',
-                )
-
-    # TODO: refactor all the name checks to be less verbose and complex
-    def _check_header_name_server_managed(
-        self,
-        responses: list[ResponseSpec],
-    ) -> None:
-        endpoint_name = self.metadata.endpoint_name
-        for response in responses:
-            if not response.headers:
-                continue
-
-            forbidden_header = self._get_forbidden_header(
-                response.headers.items(),
-            )
-
-            if forbidden_header:
-                raise EndpointMetadataError(
-                    f'Header {forbidden_header!r} is not allowed in responses '
-                    f'from endpoint {endpoint_name!r}.',
-                )
-
-    def _check_responses_header_syntax(
-        self,
-        responses: list[ResponseSpec],
-    ) -> None:
-
-        for response in responses:
-            if not response.headers:
-                continue
-
-            invalid_header = self._get_invalid_header(
-                response.headers.keys(),
-            )
-
-            if invalid_header:
-                raise EndpointMetadataError(
-                    f'Header name {invalid_header!r} is not '
-                    f'following http spec.',
-                )
-
-    def _check_modification_header_syntax(
-        self,
-    ) -> None:
-        modification = self.metadata.modification
-
-        if not modification or not modification.headers:
-            return
-
-        invalid_header = self._get_invalid_header(
-            modification.headers.keys(),
+        self.http_spec_validator(
+            metadata=self.metadata,
+        ).validate(
+            responses=responses,
         )
-
-        if invalid_header:
-            raise EndpointMetadataError(
-                f'Header name {invalid_header!r} is not following http spec.',
-            )
-
-    def _check_modification_cookie_syntax(
-        self,
-    ) -> None:
-
-        modification = self.metadata.modification
-
-        if not modification or not modification.cookies:
-            return
-
-        invalid_cookie = self._get_invalid_cookie(
-            modification.cookies.keys(),
-        )
-
-        if invalid_cookie:
-            raise EndpointMetadataError(
-                f'Cookie name {invalid_cookie!r} is not following http spec.',
-            )
-
-    def _check_response_cookie_syntax(
-        self,
-        responses: list[ResponseSpec],
-    ) -> None:
-        for response in responses:
-            if not response.cookies:
-                continue
-
-            invalid_cookie = self._get_invalid_cookie(
-                response.cookies.keys(),
-            )
-
-            if invalid_cookie:
-                raise EndpointMetadataError(
-                    f'Cookie name {invalid_cookie!r} is '
-                    f'not following http spec.',
-                )
 
     def _convert_responses(
         self,
         all_responses: list[ResponseSpec],
     ) -> dict[HTTPStatus, ResponseSpec]:
         return {resp.status_code: resp for resp in all_responses}
-
-    def _get_invalid_header(
-        self,
-        header_names: KeysView[str],
-    ) -> str | None:
-        for header_name in header_names:
-            if not _ALLOWED_TOKENS_PATTERN.match(header_name):
-                return header_name
-
-        return None
-
-    def _get_invalid_cookie(
-        self,
-        cookie_names: KeysView[str],
-    ) -> str | None:
-        for cookie_name in cookie_names:
-            if not _ALLOWED_TOKENS_PATTERN.match(cookie_name):
-                return cookie_name
-
-        return None
-
-    def _get_forbidden_header(
-        self,
-        response_headers: ItemsView[str, HeaderSpec],
-    ) -> str | None:
-
-        for header_name, header in response_headers:
-            if (
-                header_name.lower() in _FORBIDDEN_RESPONSE_HEADERS
-                and not header.skip_validation
-            ):
-                return header_name
-        return None
 
 
 @dataclasses.dataclass(slots=True, frozen=True, kw_only=True)